feat(use-case): add validations on update avatar

+++ src/domain/use-cases/user/update-avatar.ts
@@ -5,6 +5,7 @@ import {
   IUpdateUserAvatarUseCase,
   Input,
 } from "@domain/contracts/use-cases/user/update-avatar";
+import { AppError } from "@domain/errors/app-error";
 import { deleteFile } from "@domain/helpers/file";
 
 @injectable()
@@ -13,16 +14,22 @@ class UpdateUserAvatarUseCase implements IUpdateUserAvatarUseCase {
     @inject("UserRepository") private userRepository: IUserRepository
   ) {}
 
-  async execute({ user, avatar }: Input) {
-    const oldAvatar = user.avatar;
+  async execute({ user: userToUpdate, avatar }: Input) {
+    if (!avatar) {
+      throw new AppError("Avatar does not provided");
+    }
 
-    user.avatar = avatar;
+    const oldAvatar = userToUpdate.avatar;
 
-    await this.userRepository.update(user);
+    userToUpdate.avatar = avatar;
+
+    const user = await this.userRepository.update(userToUpdate);
 
     if (oldAvatar) {
       deleteFile(`./storage/avatars/${oldAvatar}`);
     }
+
+    return user;
   }
 }
